+++ app/test/frontend/utils.dart
@@ -53,7 +53,6 @@ Package createTestPackage({List<AuthenticatedUser> uploaders}) {
     ..created = new DateTime.utc(2014)
     ..updated = new DateTime.utc(2015)
     ..uploaders = uploaders.map((user) => user.userId).toList()
-    ..uploaderEmails = uploaders.map((user) => user.email).toList()
     ..latestVersionKey = testPackageVersionKey
     ..latestDevVersionKey = testPackageVersionKey;
 }