+++ CHANGELOG.md
@@ -3,6 +3,8 @@ AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
 
+ * `Package.uploaderEmails` and `PackageVersion.uploaderEmail` is no longer used/updated.
+
 ## `20190325t131912-all`
 
  * Fixes to invitation logic.